max_sub_array brute forced and minor house keeping

+++ Delve/include/Sort.hpp
@@ -6,7 +6,7 @@ namespace Sort
 template<typename T, size_t N>
 void insertion_sort(T (&array)[N]) noexcept;
 template<typename T>
-void merge_sort(T* array, int Left, int Right) noexcept;
+void merge_sort(T* array, size_t Left, size_t Right) noexcept;
 } // namespace Sort
 } // namespace Delve
